Add a Redis Adapter (#11)
This commit sets up the Redis docker-compose service and implements the cache interface. It also renames the `ci.yml` Golang workflow file to `go.yml` and refactors debug logging messages a bit.

The adapter implementation connects to the Redis container and performs basic `Get`, `Set`, `Delete` operations, using the [go-redis](https://github.com/redis/go-redis) client. It also supports working with JSON.

Redis is planned to be used in the future to store JWT sessions, as well as whitelists and blacklists of access and refresh tokens, respectively.